[Fix] test-harness: Set MRAC to same value as ROM.
This disables side effects in memory regions the CPU loads/stores from,
which is necessary for VeeR to handle unaligned loads correctly (as
enabled in 1c11a80f327c5a07be88390f648d13bca94ef147).

Because this makes the CPU run a bit faster, I also had to adjust the
test assertions around where the PC is when an bus-fault NMI happens.